requirements.txt
# AutoStream Agent — Python Dependencies
# Install with: pip install -r requirements.txt

# Core LangChain + LangGraph
langchain==0.3.25
langchain-core==0.3.58
langchain-google-genai
langgraph==0.3.34

# LLM Provider
anthropic>=0.49.0

# Environment variable management
python-dotenv>=1.0.0

# Type extensions (for TypedDict compatibility < Python 3.12)
typing-extensions>=4.9.0
